backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9